// src/graphql/__generated__/graphqlgen.ts
type Role = 'ADMIN' | 'EDITOR' | 'USER' | 'UNAUTHORIZED';
// ...
  export const defaultResolvers = {};

  export interface ArgsUser {
    id: string;
  }

  export interface ArgsDocument {
    id: string;
  }

  export interface ArgsDocumentType {
    id: string;
  }
// ...
  export interface Type {
    session: (
      parent: {},
      args: {},
      ctx: Context,
      info: GraphQLResolveInfo,
    ) => Session | null | Promise<Session | null>;

    files: (
      parent: {},
      args: {},
      ctx: Context,
      info: GraphQLResolveInfo,
    ) => File[] | Promise<File[]>;

    user: (
      parent: {},
      args: ArgsUser,
      ctx: Context,
      info: GraphQLResolveInfo,
    ) => User | null | Promise<User | null>;

    users: (
      parent: {},
      args: {},
      ctx: Context,
      info: GraphQLResolveInfo,
    ) => User[] | Promise<User[]>;

    document: (
      parent: {},
      args: ArgsDocument,
      ctx: Context,
      info: GraphQLResolveInfo,
    ) => Document | null | Promise<Document | null>;

    documents: (
      parent: {},
      args: {},
      ctx: Context,
      info: GraphQLResolveInfo,
    ) => Document[] | Promise<Document[]>;

    documentTypes: (
      parent: {},
      args: {},
      ctx: Context,
      info: GraphQLResolveInfo,
    ) => DocumentType[] | Promise<DocumentType[]>;

    documentType: (
      parent: {},
      args: ArgsDocumentType,
      ctx: Context,
      info: GraphQLResolveInfo,
    ) => DocumentType | null | Promise<DocumentType | null>;
  }
// ...
  export const defaultResolvers = {
    iat: (parent: Session) => parent.iat,
    user: (parent: Session) => parent.user,
  };
// ...
export namespace MutationResolvers {
  export const defaultResolvers = {};

  export interface LoginInput {
    email: string;
    password: string;
  }
  export interface UserCreateInput {
    email: string;
    password: string;
    repeatPassword: string;
    firstName: string;
    lastName: string;
    role: Role;
    avatar: string | null;
  }
  export interface UserUpdateInput {
    email: string;
    firstName: string;
    lastName: string;
    password: string | null;
    repeatPassword: string | null;
    role: Role;
    avatar: string | null;
  }
  export interface DocumentCreateInput {
    title: string;
    publishDate: string;
    type: string;
    author: string;
  }
  export interface DocumentUpdateInput {
    title: string;
    publishDate: string;
    type: string;
    author: string;
  }

  export interface ArgsLogin {
    input: LoginInput;
  }

  export interface ArgsUploadFile {
    file: string;
  }

  export interface ArgsDeleteFile {
    id: string;
  }

  export interface ArgsCreateUser {
    input: UserCreateInput;
  }

  export interface ArgsDeleteUser {
    id: string;
  }

  export interface ArgsUpdateUser {
    input: UserUpdateInput;
    id: string;
  }

  export interface ArgsCreateDocument {
    input: DocumentCreateInput;
  }

  export interface ArgsUpdateDocument {
    input: DocumentUpdateInput;
    id: string;
  }

  export type LoginResolver = (
    parent: {},
    args: ArgsLogin,
    ctx: Context,
    info: GraphQLResolveInfo,
  ) => string | null | Promise<string | null>;

  export type UploadFileResolver = (
    parent: {},
    args: ArgsUploadFile,
    ctx: Context,
    info: GraphQLResolveInfo,
  ) => File | Promise<File>;

  export type DeleteFileResolver = (
    parent: {},
    args: ArgsDeleteFile,
    ctx: Context,
    info: GraphQLResolveInfo,
  ) => File | null | Promise<File | null>;

  export type CreateUserResolver = (
    parent: {},
    args: ArgsCreateUser,
    ctx: Context,
    info: GraphQLResolveInfo,
  ) => User | Promise<User>;

  export type DeleteUserResolver = (
    parent: {},
    args: ArgsDeleteUser,
    ctx: Context,
    info: GraphQLResolveInfo,
  ) => User | null | Promise<User | null>;

  export type UpdateUserResolver = (
    parent: {},
    args: ArgsUpdateUser,
    ctx: Context,
    info: GraphQLResolveInfo,
  ) => User | null | Promise<User | null>;

  export type CreateDocumentResolver = (
    parent: {},
    args: ArgsCreateDocument,
    ctx: Context,
    info: GraphQLResolveInfo,
  ) => Document | Promise<Document>;

  export type UpdateDocumentResolver = (
    parent: {},
    args: ArgsUpdateDocument,
    ctx: Context,
    info: GraphQLResolveInfo,
  ) => Document | null | Promise<Document | null>;

  export interface Type {
    login: (
      parent: {},
      args: ArgsLogin,
      ctx: Context,
      info: GraphQLResolveInfo,
    ) => string | null | Promise<string | null>;

    uploadFile: (
      parent: {},
      args: ArgsUploadFile,
      ctx: Context,
      info: GraphQLResolveInfo,
    ) => File | Promise<File>;

    deleteFile: (
      parent: {},
      args: ArgsDeleteFile,
      ctx: Context,
      info: GraphQLResolveInfo,
    ) => File | null | Promise<File | null>;

    createUser: (
      parent: {},
      args: ArgsCreateUser,
      ctx: Context,
      info: GraphQLResolveInfo,
    ) => User | Promise<User>;

    deleteUser: (
      parent: {},
      args: ArgsDeleteUser,
      ctx: Context,
      info: GraphQLResolveInfo,
    ) => User | null | Promise<User | null>;

    updateUser: (
      parent: {},
      args: ArgsUpdateUser,
      ctx: Context,
      info: GraphQLResolveInfo,
    ) => User | null | Promise<User | null>;

    createDocument: (
      parent: {},
      args: ArgsCreateDocument,
      ctx: Context,
      info: GraphQLResolveInfo,
    ) => Document | Promise<Document>;

    updateDocument: (
      parent: {},
      args: ArgsUpdateDocument,
      ctx: Context,
      info: GraphQLResolveInfo,
    ) => Document | null | Promise<Document | null>;
  }
}